refactor : progressHistory progress enum 적용
