+++ Whatnow-Api/src/main/kotlin/com/depromeet/whatnow/api/promiseprogress/dto/response/PromiseProgressDto.kt
@@ -1,7 +1,7 @@
 package com.depromeet.whatnow.api.promiseprogress.dto.response
 
-import com.depromeet.whatnow.domains.promiseprogress.domain.PromiseProgress
-import com.depromeet.whatnow.domains.promiseprogress.domain.PromiseProgressGroup
+import com.depromeet.whatnow.domains.progresshistory.domain.PromiseProgress
+import com.depromeet.whatnow.domains.progresshistory.domain.PromiseProgressGroup
 
 data class PromiseProgressDto(
     val group: PromiseProgressGroup,
