+++ Whatnow-Api/src/main/kotlin/com/depromeet/whatnow/api/promiseprogress/dto/response/PromiseProgressGroupElement.kt
@@ -1,6 +1,6 @@
 package com.depromeet.whatnow.api.promiseprogress.dto.response
 
-import com.depromeet.whatnow.domains.promiseprogress.domain.PromiseProgressGroup
+import com.depromeet.whatnow.domains.progresshistory.domain.PromiseProgressGroup
 
 data class PromiseProgressGroupElement(
     val group: PromiseProgressGroup,
