+++ Whatnow-Api/src/main/kotlin/com/depromeet/whatnow/api/promiseprogress/usecase/PromiseProgressReadUseCase.kt
@@ -3,7 +3,6 @@ package com.depromeet.whatnow.api.promiseprogress.usecase
 import com.depromeet.whatnow.annotation.UseCase
 import com.depromeet.whatnow.api.promiseprogress.dto.response.PromiseProgressDto
 import com.depromeet.whatnow.api.promiseprogress.dto.response.PromiseProgressGroupElement
-import com.depromeet.whatnow.domains.promiseprogress.adapter.PromiseProgressAdapter
 
 @UseCase
 class PromiseProgressReadUseCase(
