+++ Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/domains/progresshistory/domain/ProgressHistory.kt
@@ -1,7 +1,6 @@
 package com.depromeet.whatnow.domains.progresshistory.domain
 
-import com.depromeet.whatnow.consts.DEFAULT_PROMISE_PROGRESS
-import com.depromeet.whatnow.domains.progresshistory.exception.PromiseIsSameException
+import com.depromeet.whatnow.domains.progresshistory.exception.ProgressIsSameException
 import javax.persistence.Column
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
@@ -17,28 +16,26 @@ class ProgressHistory(
 
     var userId: Long,
 
-    var currentPromiseProgressId: Long?,
+    var currentPromiseProgress: PromiseProgress,
 
-    var prePromiseProgressId: Long?,
+    var prePromiseProgress: PromiseProgress?,
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "progress_history_id")
     val id: Long? = null,
 ) {
-    fun change(progressId: Long) {
-        if (progressId == currentPromiseProgressId) {
-            throw PromiseIsSameException.EXCEPTION
+    fun change(progress : PromiseProgress) {
+        if (progress == currentPromiseProgress) {
+            throw ProgressIsSameException.EXCEPTION
         }
-        currentPromiseProgressId?.let { // init default 이지 않을 때
-            prePromiseProgressId = currentPromiseProgressId
-        }
-        currentPromiseProgressId = progressId
+        prePromiseProgress = currentPromiseProgress
+        currentPromiseProgress = progress
     }
 
     companion object {
         fun of(progressId: Long, userId: Long): ProgressHistory {
-            return ProgressHistory(progressId, userId, DEFAULT_PROMISE_PROGRESS, null)
+            return ProgressHistory(progressId, userId, PromiseProgress.DEFAULT,null)
         }
     }
 }
