+++ Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/domains/progresshistory/exception/ProgressIsSameException.kt
@@ -2,10 +2,10 @@ package com.depromeet.whatnow.domains.progresshistory.exception
 
 import com.depromeet.whatnow.exception.WhatnowCodeException
 
-class PromiseIsSameException : WhatnowCodeException(
+class ProgressIsSameException : WhatnowCodeException(
     PromiseHistoryErrorCode.PROMISE_PROGRESS_IS_SAME,
 ) {
     companion object {
-        val EXCEPTION: WhatnowCodeException = PromiseIsSameException()
+        val EXCEPTION: WhatnowCodeException = ProgressIsSameException()
     }
 }
