+++ Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/domains/promise/service/PromiseDomainService.kt
@@ -4,7 +4,7 @@ import com.depromeet.whatnow.common.vo.PlaceVo
 import com.depromeet.whatnow.domains.promise.adaptor.PromiseAdaptor
 import com.depromeet.whatnow.domains.promise.domain.Promise
 import com.depromeet.whatnow.domains.promise.domain.PromiseType
-import com.depromeet.whatnow.domains.promiseprogress.domain.PromiseProgressGroup
+import com.depromeet.whatnow.domains.progresshistory.domain.PromiseProgressGroup
 import org.springframework.stereotype.Service
 import java.time.LocalDateTime
 import javax.transaction.Transactional
