--- Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/domains/promiseprogress/exception/PromiseProgressNotFoundException.kt
@@ -1,11 +1,0 @@
-package com.depromeet.whatnow.domains.promiseprogress.exception
-
-import com.depromeet.whatnow.exception.WhatnowCodeException
-
-class PromiseProgressNotFoundException : WhatnowCodeException(
-    PromiseProgressCode.NOT_FOUND,
-) {
-    companion object {
-        val EXCEPTION: WhatnowCodeException = PromiseProgressNotFoundException()
-    }
-}
