--- Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/domains/promiseprogress/repository/PromiseProgressRepository.kt
@@ -1,8 +1,0 @@
-package com.depromeet.whatnow.domains.promiseprogress.repository
-
-import com.depromeet.whatnow.domains.promiseprogress.domain.PromiseProgress
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface PromiseProgressRepository : JpaRepository<PromiseProgress, Long> {
-    fun findByCode(code: String): PromiseProgress?
-}
